fix(graph): correct typo in user identification comment and reorganize schema imports
